refactor: Controller의 메서드 분리
